% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maparpav-package.R
\docType{package}
\name{maparpav-package}
\alias{maparpav}
\alias{maparpav-package}
\title{maparpav: Retreieve map from ARPAV}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Few times a day, ARPAV update map data on a site
    dedicated to UBEP, ie https://www.arpa.veneto.it/inquinanti/unipd/.
    Here there are the utilities and function to automatically download
    those data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/maparpav}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/maparpav/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [copyright holder]

}
\keyword{internal}
